refactored route registering
