+++ client/keys/root.go
@@ -2,6 +2,7 @@ package keys
 
 import (
 	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/gorilla/mux"
 	"github.com/spf13/cobra"
 )
 
@@ -26,4 +27,13 @@ func Commands() *cobra.Command {
 		updateKeyCommand(),
 	)
 	return cmd
+}
+
+func RegisterRoutes(r *mux.Router) {
+	r.HandleFunc("/keys", QueryKeysRequestHandler).Methods("GET")
+	r.HandleFunc("/keys", AddNewKeyRequestHandler).Methods("POST")
+	r.HandleFunc("/keys/seed", SeedRequestHandler).Methods("GET")
+	r.HandleFunc("/keys/{name}", GetKeyRequestHandler).Methods("GET")
+	r.HandleFunc("/keys/{name}", UpdateKeyRequestHandler).Methods("PUT")
+	r.HandleFunc("/keys/{name}", DeleteKeyRequestHandler).Methods("DELETE")
 }
