+++ client/rpc/root.go
@@ -1,6 +1,7 @@
 package rpc
 
 import (
+	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
 	"github.com/spf13/cobra"
 
@@ -41,4 +42,13 @@ func initClientCommand() *cobra.Command {
 	cmd.Flags().String(flagCommit, "", "File with trusted and signed header")
 	cmd.Flags().String(flagValHash, "", "Hash of trusted validator set (hex-encoded)")
 	return cmd
+}
+
+func RegisterRoutes(r *mux.Router) {
+	r.HandleFunc("/node_info", NodeInfoRequestHandler).Methods("GET")
+	r.HandleFunc("/syncing", NodeSyncingRequestHandler).Methods("GET")
+	r.HandleFunc("/blocks/latest", LatestBlockRequestHandler).Methods("GET")
+	r.HandleFunc("/blocks/{height}", BlockRequestHandler).Methods("GET")
+	r.HandleFunc("/validatorsets/latest", LatestValidatorsetRequestHandler).Methods("GET")
+	r.HandleFunc("/validatorsets/{height}", ValidatorsetRequestHandler).Methods("GET")
 }
