+++ client/tx/root.go
@@ -1,6 +1,7 @@
 package tx
 
 import (
+	"github.com/gorilla/mux"
 	"github.com/spf13/cobra"
 
 	"github.com/cosmos/cosmos-sdk/wire"
@@ -18,4 +19,11 @@ func AddCommands(cmd *cobra.Command, cdc *wire.Codec) {
 		SearchTxCmd(cmdr),
 		QueryTxCmd(cmdr),
 	)
+}
+
+func RegisterRoutes(r *mux.Router, cdc *wire.Codec) {
+	r.HandleFunc("/txs", SearchTxRequestHandler(cdc)).Methods("GET")
+	r.HandleFunc("/txs/{hash}", QueryTxRequestHandler(cdc)).Methods("GET")
+	r.HandleFunc("/txs/sign", SignTxRequstHandler).Methods("POST")
+	r.HandleFunc("/txs/broadcast", BroadcastTxRequestHandler).Methods("POST")
 }
